Changes for increased 64bit compatibility.

+++ test/test_vector.cpp
@@ -485,12 +485,12 @@ namespace
       Compare_Data compare_data;
       Data data;
 
-      for (size_t i = 0; i < SIZE; ++i)
+      for (int i = 0; i < int(SIZE); ++i)
       {
         compare_data.push_back(i);
       }
 
-      for (size_t i = 0; i < SIZE; ++i)
+      for (int i = 0; i < int(SIZE); ++i)
       {
         data.push_back(i);
       }
@@ -529,7 +529,7 @@ namespace
     {
       Data data;
 
-      for (size_t i = 0; i < SIZE; ++i)
+      for (int i = 0; i < int(SIZE); ++i)
       {
         data.push_back(i);
       }
